Excluir endereço e início de atualizar endereço

+++ app/src/main/java/com/example/khiata/models/Address.java
@@ -1,7 +1,7 @@
 package com.example.khiata.models;
 
 public class Address {
-    private long id=0;
+    private int id=0;
     private String recipient;
     private String street;
     private int number;
@@ -16,11 +16,11 @@ public class Address {
         this.label = label;
     }
 
-    public long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(int id) {
         this.id = id;
     }
